Match language headers with spaces. This is common in Gherkin docs.
This implementation uses regular expressions to match a token, which
we should generally avoid because it is slow. However, this state
is only entered once per document, so the penalty should be low.
If we need to optimise it we can check if the 1st char of the trimmed
line is a '#'.