Almost implemented Ast comments for Go.
This is my first stab at writing Go code. Nice to get started!
The code seems to work when there is at least one comment -
the comments are written to the JSON representatin.

However, when there are no comments the comments property
in the JSON is null, not [] which is what it should be.

Any clues @muhqu ?