Compiler (mostly) implemented in Java.
To simplify the design I decided to reuse classes from the Ast package.
That exposed a subtle difference - the test cases (and arguments) need
a list of locations, so reusing the same DocString and DataTable classes
is probably a bad idea.

I'll probably reintroduce the Pickles. I thought for a while that it
would be simpler to have TestCases in Gherkin, and let Cucumber simply
create derived TestCases with extra Before/After hooks, and also
"linking" step definitions.

It's become clear to me that it's better to differentiate between those
bounded contexts: compilation and execution.

Next step will be to reintroduce pickles. Cucumber can create TestCases
from those. They'll be similar, but slightly different (they might have
slots for linking stepdefs).