Build objective-c on Travis. Fix Travis language settings.
Build (including the acceptance tests) the objective-c version on the
top Travis build. Compile the objective-c version on the
gherkin-objective-c Travis build.

Set the language explicitly on all jobs in the Travis matrix, otherwise
the default version of the language is used instead of the specified
one.

Use jruby 1.7.22 instead of jruby-head, the jruby-head version issues
warnings that interfere with the error tests.

Run the unit test for the Python version on the top Travis build.